// site/convex/_generated/api.ts
/* eslint-disable */
/**
 * Generated `api` object type definitions.
 *
 * THIS FILE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.34.1.
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
  OptionalRestArgs,
} from "convex/server";
import type * as logs from "../logs";
import type * as mentions from "../mentions";
import type * as signals from "../signals";

/**
 * A type describing your Convex backend's API.
 *
 * This `API` type includes information about the arguments and return
 * types of your Convex functions. It is used to enable type checking when
 * using the Convex client SDK.
 *
 * @public
 */
export type API = ApiFromModules<{
  logs: typeof logs;
  mentions: typeof mentions;
  signals: typeof signals;
}>;

export const api = new Proxy(
  {},
  {
    get:
      (_target, prop: string | symbol) =>
      (...args: unknown[]) => {
        throw new Error(
          `API function called on the server. ` +
            `Use "convex/server" instead of "convex/client" for server code.`
        );
      },
  }
) as never;